+++ pkg/tests/end_to_end_tests/ha_single_promscale_test.go
@@ -125,7 +125,7 @@ func prepareWriterWithHa(db *pgxpool.Pool, t testing.TB) (*util.ManualTicker, ht
 	mCache := &cache.MetricNameCache{Metrics: clockcache.WithMax(cache.DefaultMetricCacheSize)}
 
 	ing, err := ingestor.NewPgxIngestor(pgxconn.NewPgxConn(db), mCache, sCache, nil, &ingestor.Cfg{
-		InvertedLabelsCacheSize: cache.DefaultConfig.InvertedLabelsCacheSize,
+		InvertedLabelsCacheSize: cache.DefaultConfig.InvertedLabelsCacheSize, NumCopiers: 2,
 	})
 	if err != nil {
 		t.Fatalf("could not create ingestor: %v", err)